refactor(health): extract continuous health monitor predicate to meshtastic_utils
Extract the `_requires_continuous_health_monitor()` logic from main.py into a
new `meshtastic_utils.requires_continuous_health_monitor()` function to avoid
code duplication. The new function also adds early exit logic in
`check_connection()` to skip health monitoring setup entirely when not needed
for the connection type.

Additionally:
- Pin aiohttp to exact version 3.13.3 in requirements.txt and setup.py
- Add test helper `reset_meshtastic_utils_globals()` in tests/helpers.py
- Refactor test_db_utils.py to extract `_make_failing_cursor_proxy_side_effect()`
- Update test_main.py to use the shared test helper function
