+++ requirements.txt
@@ -1,5 +1,6 @@
 meshtastic>=2.7.8
 Pillow==12.1.1
+aiohttp==3.13.3
 matrix-nio==0.25.2
 matplotlib==3.10.8
 requests==2.32.5
