+++ tests/helpers.py
@@ -6,6 +6,8 @@ and improve maintainability of the test suite.
 """
 
 import asyncio
+import contextlib
+import sys
 from typing import Any, Callable, TypeVar, cast
 
 T = TypeVar("T")
@@ -123,3 +125,106 @@ def make_patched_get_running_loop() -> Callable[[], asyncio.AbstractEventLoop]:
         return cast(asyncio.AbstractEventLoop, wrapped)
 
     return _patched_get_running_loop
+
+
+def reset_meshtastic_utils_globals(*, shutdown_executors: bool = False) -> None:
+    """
+    Reset meshtastic_utils globals shared across tests.
+
+    This function resets module-level global state in mmrelay.meshtastic_utils
+    to clean defaults, optionally shutting down thread pool executors.
+
+    Parameters:
+        shutdown_executors: If True, shut down _metadata_executor and _ble_executor
+            before resetting them to None.
+    """
+    if "mmrelay.meshtastic_utils" not in sys.modules:
+        return
+
+    module = sys.modules["mmrelay.meshtastic_utils"]
+    if hasattr(module, "config"):
+        module.config = None  # type: ignore[attr-defined]
+    if hasattr(module, "matrix_rooms"):
+        module.matrix_rooms = []  # type: ignore[attr-defined]
+    if hasattr(module, "meshtastic_client"):
+        module.meshtastic_client = None  # type: ignore[attr-defined]
+    if hasattr(module, "meshtastic_iface"):
+        module.meshtastic_iface = None  # type: ignore[attr-defined]
+    if hasattr(module, "event_loop"):
+        module.event_loop = None  # type: ignore[attr-defined]
+    if hasattr(module, "reconnecting"):
+        module.reconnecting = False  # type: ignore[attr-defined]
+    if hasattr(module, "shutting_down"):
+        module.shutting_down = False  # type: ignore[attr-defined]
+    if hasattr(module, "reconnect_task"):
+        module.reconnect_task = None  # type: ignore[attr-defined]
+    if hasattr(module, "subscribed_to_messages"):
+        module.subscribed_to_messages = False  # type: ignore[attr-defined]
+    if hasattr(module, "subscribed_to_connection_lost"):
+        module.subscribed_to_connection_lost = False  # type: ignore[attr-defined]
+    if hasattr(module, "_metadata_future"):
+        module._metadata_future = None  # type: ignore[attr-defined]
+    if hasattr(module, "_metadata_future_started_at"):
+        module._metadata_future_started_at = None  # type: ignore[attr-defined]
+    if hasattr(module, "_ble_future"):
+        module._ble_future = None  # type: ignore[attr-defined]
+    if hasattr(module, "_ble_future_address"):
+        module._ble_future_address = None  # type: ignore[attr-defined]
+    if hasattr(module, "_ble_future_started_at"):
+        module._ble_future_started_at = None  # type: ignore[attr-defined]
+    if hasattr(module, "_ble_future_timeout_secs"):
+        module._ble_future_timeout_secs = None  # type: ignore[attr-defined]
+    if hasattr(module, "_ble_timeout_counts"):
+        module._ble_timeout_counts = {}  # type: ignore[attr-defined]
+    if hasattr(module, "_ble_executor_orphaned_workers_by_address"):
+        module._ble_executor_orphaned_workers_by_address = {}  # type: ignore[attr-defined]
+    if hasattr(module, "_metadata_executor_orphaned_workers"):
+        module._metadata_executor_orphaned_workers = 0  # type: ignore[attr-defined]
+    if hasattr(module, "_ble_executor_degraded_addresses"):
+        module._ble_executor_degraded_addresses = set()  # type: ignore[attr-defined]
+    if hasattr(module, "_metadata_executor_degraded"):
+        module._metadata_executor_degraded = False  # type: ignore[attr-defined]
+    if hasattr(module, "_health_probe_request_deadlines"):
+        module._health_probe_request_deadlines = {}  # type: ignore[attr-defined]
+    if hasattr(module, "_ble_future_watchdog_secs"):
+        module._ble_future_watchdog_secs = getattr(  # type: ignore[attr-defined]
+            module, "BLE_FUTURE_WATCHDOG_SECS", module._ble_future_watchdog_secs
+        )
+    if hasattr(module, "_ble_timeout_reset_threshold"):
+        module._ble_timeout_reset_threshold = getattr(  # type: ignore[attr-defined]
+            module, "BLE_TIMEOUT_RESET_THRESHOLD", module._ble_timeout_reset_threshold
+        )
+    if hasattr(module, "_ble_scan_timeout_secs"):
+        module._ble_scan_timeout_secs = getattr(  # type: ignore[attr-defined]
+            module, "BLE_SCAN_TIMEOUT_SECS", module._ble_scan_timeout_secs
+        )
+    if hasattr(module, "_ble_future_stale_grace_secs"):
+        module._ble_future_stale_grace_secs = getattr(  # type: ignore[attr-defined]
+            module, "BLE_FUTURE_STALE_GRACE_SECS", module._ble_future_stale_grace_secs
+        )
+    if hasattr(module, "_ble_interface_create_timeout_secs"):
+        module._ble_interface_create_timeout_secs = getattr(  # type: ignore[attr-defined]
+            module,
+            "BLE_INTERFACE_CREATE_TIMEOUT_FLOOR_SECS",
+            module._ble_interface_create_timeout_secs,
+        )
+
+    def _shutdown_executor(executor: Any) -> None:
+        try:
+            executor.shutdown(wait=False, cancel_futures=True)
+        except TypeError:
+            with contextlib.suppress(RuntimeError):
+                executor.shutdown(wait=False)
+        except RuntimeError:
+            pass
+
+    if shutdown_executors and hasattr(module, "_metadata_executor"):
+        executor = module._metadata_executor  # type: ignore[attr-defined]
+        if executor is not None:
+            _shutdown_executor(executor)
+        module._metadata_executor = None  # type: ignore[attr-defined]
+    if shutdown_executors and hasattr(module, "_ble_executor"):
+        executor = module._ble_executor  # type: ignore[attr-defined]
+        if executor is not None:
+            _shutdown_executor(executor)
+        module._ble_executor = None  # type: ignore[attr-defined]
